+++ common/src/main/java/net/mehvahdjukaar/vista/client/TvBlockEntityRenderer.java
@@ -11,6 +11,7 @@ import net.mehvahdjukaar.vista.common.CassetteTape;
 import net.mehvahdjukaar.vista.common.TVBlock;
 import net.mehvahdjukaar.vista.common.TVBlockEntity;
 import net.mehvahdjukaar.vista.configs.ClientConfigs;
+import net.mehvahdjukaar.vista.integration.CompatHandler;
 import net.mehvahdjukaar.vista.integration.exposure.ExposureCompatClient;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
@@ -19,13 +20,11 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
-import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.resources.model.Material;
 import net.minecraft.core.Direction;
 import net.minecraft.core.Holder;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.Mth;
 import net.minecraft.world.item.ItemStack;
 
 import java.util.UUID;
@@ -51,7 +50,7 @@ public class TvBlockEntityRenderer implements BlockEntityRenderer<TVBlockEntity>
         LOD lod = LOD.at(blockEntity);
         //TODO: change with tv size
 
-        if(!lod.isMedium())return;
+        if (!lod.isMedium()) return;
 
         if (lod.isPlaneCulled(dir, 0.5f, 0f)) {
             return;
@@ -94,7 +93,7 @@ public class TvBlockEntityRenderer implements BlockEntityRenderer<TVBlockEntity>
 
                 vc = mat.buffer(buffer, t -> ModRenderTypes.CAMERA_DRAW_SPRITE.apply(t, mat));
             }
-        } else if (VistaMod.EXPOSURE_ON) {
+        } else if (CompatHandler.EXPOSURE) {
             ResourceLocation texture = ExposureCompatClient.getPictureTextureForRenderer(stack, blockEntity.getAnimationTick());
             if (texture != null) {
                 vc = TapeTextureManager.getFullSpriteVC(texture, buffer, drawingCamera, 0);
@@ -119,43 +118,40 @@ public class TvBlockEntityRenderer implements BlockEntityRenderer<TVBlockEntity>
     private void maybeRenderDebug(ResourceLocation tex, PoseStack poseStack, MultiBufferSource buffer, float partialTick,
                                   TVBlockEntity tile) {
 
-        if (!ClientConfigs.isDebugOn() || !Minecraft.getInstance().getEntityRenderDispatcher().shouldRenderHitBoxes()) return;
+        if (!ClientConfigs.isDebugOn() || !Minecraft.getInstance().getEntityRenderDispatcher().shouldRenderHitBoxes())
+            return;
         poseStack.pushPose();
 
-        try {
-            Font font = Minecraft.getInstance().font;
+        Font font = Minecraft.getInstance().font;
 
-            poseStack.translate(1, 1.5, 0);
-            poseStack.mulPose(Axis.YP.rotationDegrees(-180));
-            poseStack.scale(1 / 16f, -1 / 16f, 1 / 16f);
-            poseStack.scale(0.5f, 0.5f, 0.5f);
+        poseStack.translate(1, 1.5, 0);
+        poseStack.mulPose(Axis.YP.rotationDegrees(-180));
+        poseStack.scale(1 / 16f, -1 / 16f, 1 / 16f);
+        poseStack.scale(0.5f, 0.5f, 0.5f);
 
-            RollingBuffer<Long> lastUpdateTimes = LiveFeedRendererManager.UPDATE_TIMES.get(tex);
+        RollingBuffer<Long> lastUpdateTimes = LiveFeedRendererManager.UPDATE_TIMES.get(tex);
 
-            double averageUpdateinterval = calculateAverageUpdateTime(lastUpdateTimes);
+        double averageUpdateinterval = calculateAverageUpdateTime(lastUpdateTimes);
 
-            int y = 0;
-            font.drawInBatch(String.format("up rate %.2f", averageUpdateinterval), 0, 7, -1,
-                    false, poseStack.last().pose(), buffer, Font.DisplayMode.NORMAL,
-                    OverlayTexture.NO_OVERLAY,
-                    LightTexture.FULL_BRIGHT);
+        int y = 0;
+        font.drawInBatch(String.format("up rate %.2f", averageUpdateinterval), 0, 7, -1,
+                false, poseStack.last().pose(), buffer, Font.DisplayMode.NORMAL,
+                OverlayTexture.NO_OVERLAY,
+                LightTexture.FULL_BRIGHT);
 
-            double updateMs = LiveFeedRendererManager.SCHEDULER.get().getAverageUpdateTimeMs();
+        double updateMs = LiveFeedRendererManager.SCHEDULER.get().getAverageUpdateTimeMs();
 
+        y -= 9;
+        font.drawInBatch(String.format("up ms %.2f", updateMs), 0, y, -1, false, poseStack.last().pose(), buffer, Font.DisplayMode.NORMAL,
+                OverlayTexture.NO_OVERLAY, LightTexture.FULL_BRIGHT);
+
+        float endermanAnim = tile.getLookingAtEndermanAnimation(partialTick);
+        if (endermanAnim != 0) {
             y -= 9;
-            font.drawInBatch(String.format("up ms %.2f", updateMs), 0, y, -1, false, poseStack.last().pose(), buffer, Font.DisplayMode.NORMAL,
+
+            font.drawInBatch("p " + endermanAnim, 0, y, -1, false, poseStack.last().pose(), buffer, Font.DisplayMode.NORMAL,
                     OverlayTexture.NO_OVERLAY, LightTexture.FULL_BRIGHT);
 
-            float endermanAnim = tile.getLookingAtEndermanAnimation(partialTick);
-            if (endermanAnim != 0) {
-                y -= 9;
-
-                font.drawInBatch("p " + endermanAnim, 0, y, -1, false, poseStack.last().pose(), buffer, Font.DisplayMode.NORMAL,
-                        OverlayTexture.NO_OVERLAY, LightTexture.FULL_BRIGHT);
-
-            }
-        } catch (Exception ignored) {
-            int aa = 1;
         }
         poseStack.popPose();
     }