+++ common/src/main/java/net/mehvahdjukaar/vista/client/ViewFinderController.java
@@ -146,7 +146,7 @@ public class ViewFinderController {
 
         if (scrollDelta != 0) {
             ViewFinderBlockEntity tile = access.getInternalTile();
-            int newZoom = (Math.clamp((int) (tile.getZoomLevel() + scrollDelta), 1, tile.getMaxZoom()));
+            int newZoom = (Math.clamp((int) (tile.getZoomLevel() + scrollDelta), 1, ViewFinderBlockEntity.MAX_ZOOM));
             int oldZoom = tile.getZoomLevel();
             if (newZoom != oldZoom) {
                 tile.setZoomLevel(newZoom);