+++ common/src/main/java/net/mehvahdjukaar/vista/configs/ClientConfigs.java
@@ -5,6 +5,7 @@ import net.mehvahdjukaar.moonlight.api.platform.configs.ConfigBuilder;
 import net.mehvahdjukaar.moonlight.api.platform.configs.ConfigType;
 import net.mehvahdjukaar.moonlight.api.platform.configs.ModConfigHolder;
 import net.mehvahdjukaar.vista.VistaMod;
+import net.mehvahdjukaar.vista.integration.CompatHandler;
 
 import java.util.function.Supplier;
 
@@ -57,6 +58,8 @@ public class ClientConfigs {
         RENDER_DEBUG = builder
                 .comment("Enables rendering of debug information for televisions")
                 .define("render_debug", false);
+
+        CompatHandler.addConfigs(builder);
 
         builder.pop();
         builder.pop();